+++ app/src/main/AndroidManifest.xml
@@ -30,10 +30,6 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/WearMmrTheme">
-        <activity
-            android:name=".ui.MainComposeActivity"
-            android:exported="false"
-            android:label="@string/title_activity_main_compose" />
 
         <uses-library
             android:name="com.google.android.wearable"
