--- app/src/main/res/layout/activity_main.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-  ~    Copyright 2022 Andrey Mukamolov
-  ~
-  ~    Licensed under the Apache License, Version 2.0 (the "License");
-  ~    you may not use this file except in compliance with the License.
-  ~    You may obtain a copy of the License at
-  ~
-  ~        http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~    Unless required by applicable law or agreed to in writing, software
-  ~    distributed under the License is distributed on an "AS IS" BASIS,
-  ~    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~    See the License for the specific language governing permissions and
-  ~    limitations under the License.
-  -->
-
-<androidx.wear.widget.drawer.WearableDrawerLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="@dimen/box_inset_layout_padding"
-    tools:context="io.github.fobo66.wearmmr.ui.SettingsActivity"
-    tools:deviceIds="wear"
-    >
-  <androidx.core.widget.NestedScrollView
-      android:id="@+id/content_wrapper"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      android:fillViewport="true"
-      >
-
-    <include
-        android:id="@+id/content"
-        layout="@layout/content_main" />
-
-  </androidx.core.widget.NestedScrollView>
-  <androidx.wear.widget.drawer.WearableActionDrawerView
-      android:id="@+id/bottom_action_drawer"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      app:actionMenu="@menu/action_drawer_menu"
-      />
-</androidx.wear.widget.drawer.WearableDrawerLayout>
